Enable loading of 32-bit float WAV files by switching from pydub to scipy.io.wavread
This edit enable loading of 32-bit float WAV files by switching from
pydub to scipy.io.wavread

It is related to issue #328.
The existing function `read_audio_generic` relies on
`pydyb.AudioSegment` to load the data, but has a few problems:
1. Assumes that data is int16 or int32,
2. There no obvious (at least to me) way to query if the underlying
wav-file data is actually 32-bit float
3. Relies on deprecated `numpy.fromstring` conversion, which should be
replaced by `numpy.frombuffer`

For ".wav" files, the problem is solved by using `scipy.io.wavread` in
`read_audio_file`, instead of calling `read_audio_generic` and `pydyb`.
For ".au" files, the problem of loading 32-bit float audio remains, but
they are relatively rarely used.
As for the ".mp3" and ".ogg" files, they are non-PCM, lossy compression
formats, and do not have associated bit depths[*], so no change is
needed.

[*] Bit depth is only meaningful in reference to a PCM digital signal.
Ogg Vorbis Decoder has the option to set the output to a floating
number, which is not the same thing as bit depth of PCM-encoded data.